fix(model-elements)!: Fix duplication of dynamically typed work items
There are serializers that change the Capella type
(e.g. `SystemComponent` -> `SystemActor`) based on attributes
(`is_actor`). If these types are not appearing in the config, then no
work items of this type won't be fetched via the REST API and in return
they are created as new work items. This commit filters work items with
invalid/unconfigured types out and logs these types.