ci: Avoid dirtying the worktree
Patching the pyproject.toml file like before led to a dirty worktree,
which caused setuptools-scm to always generate dev versions even when
making wheels for a tag.

This commit refactors the CI so that patching the pyproject.toml file is
no longer necessary. The (yet) private dependency project will be
install directly from the Github repository as additional step before
installing the rest of the test dependencies, and the Gitlab pipelines
don't need to do anything special because they pull from an internal
repository which hosts the dependency package.