feat: Configure links
This commit enables a more indepth configuration of work item links. Now
the Polarion ID and capellambse attribute name are separate which allows
for a more rigorous configuration on Polarion. Possibly multiple link
configurations for the same capellambse attribute on different types.

On top there is an `include` parameter. With this parameter Additional
information can be included into the grouped link custom fields produced
for work item links. This feature is for making information from link
targets available on the link source with clickable links in the
generated HTML list.