fix: Show affected UUID in "element not found" warnings
When preparing links in markup for Polarion, but the target element
either doesn't exist in the model or has no work item associated, show
the target's UUID in the emitted warning messages. While the ID of the
source element is already logged automatically, adding the target ID
makes it easier and faster to figure out whether we have a bug or a
legitimate dead link which a model engineer needs to fix.